tablegen: Move LlvmTypeBuilder into its own file
While we're at it, make the handling of Attr constraints more robust by
mapping them to the underlying Type *before* looking them up in the
map of types that have already been built.
